feat: Add concept history tracking
Introduces a new `ExplainedConcept` type and integrates it into the `Document` interface. This allows the application to store and display a history of concepts that have been explained to the user for a given document.

Also adds a new 'concepts' tab to the `Workspace` component and updates relevant icons and services to support this new feature.